setup.py - add trove classifiers and version parsing
Found that `prompt_toolkit` is not listed as Python 3 compatible by `caniusepython3`, because it's lacking Trove Classifiers. Added them. Due to adding a License Trove Classifier, it's no longer necessary to provide the `license` argument to `setup`.

Whilst I was at it I saw you have the version defined in two files with a comment to remind you about updating both. I have adapted some code I use on a number of projects to ease this burden so the version is only stored in `__init__`.

Conflicts:
	setup.py