Big refactoring of the Window and user control interfaces. Huge performance improvements for big inputs.
(While first, it was doable to edit a file with about 1000 lines, now we can
edit 100,000 lines on the same system.)

The Window and UIControl have been rewritten very much. Rather than each time
rendering the whole user control, we now only have to render the visible part.
(Saving a lot of CPU.)

Because of this, many pieces had to be rewritten:
- UIControls work differently. They return a UIContent instance that consist
  of a collection of lines.
- All processors have been rewritten.
- Lexers work differently. `Lexer.lex_document` should now return a function
  that returns the tokens for one line. PygmentsLexer has been optimized that
  it becomes 'lazy', and it has optional syntax synchronisation. That means,
  that the lexer doesn't have to start the lexing at the beginning of the document.
  (Which would be slow for big documents.)
- Some bugs have been fixed.
- The document class got many performance improvements.