Set VMIN to 1 when entering raw mode
VMIN defines the number of characters read at a time in non-canonical
mode. It seems to default to 1 on Linux, but on Solaris and derived
operating systems it defaults to 4. (This is because the VMIN slot is
the same as the VEOF slot, which defaults to ASCII EOT = Ctrl-D = 4.)

(Thanks to @gmcmanus)