Also accept callables for the 'style' arguments for containers. Removed 'focussed' class in 'Window'.
For the styling, the 'focus' concept usually doesn't apply on a single
'Window', but on a set of containers. Adding this by default on the
`Window` doesn't make sense. If needed, we better create a helper
container to insert the 'focus' style to all its children if one of the
children is focussed.