Respect \pset pager on expected behavior
"\pset pager" has three possible values: "always", "on", and "off".
pgcli previously treated all non-"off" values as "always". This change
implements the expected behavior, which is to use the pager when the
output is larger than the terminal height (See \pset pager in
https://www.postgresql.org/docs/9.2/static/app-psql.html). Pgcli adds to
this by also using the pager when the output is wider than the terminal
width.

Fixes #813