pgcli bash completion script
This file is a bash completion script for pgcli.
It allows to <tab> complete database names, users and hosts, plus showing the other available options when calling pgcli atthe terminal prompt.
In order to work it steals a couple of functions from psql, which is the postgres CLI client installed by default. 
It is intended to be packaged with the debian package and to be copied on /usr/share/bash-completion/completions/pgcli (filename changes for consistency) upon installation of the package.
Please make the debian package mantainer aware of it and include it in the relevant section of the installation scripts/package.
I hope it is useful.