setup.py: use wheel by default, latest psycopg2
Motivation:
* use the The psycopg2 wheel (pre-compiled binary) for convenience where possible.
* allow use of psycopg2.8 (and greater!) now that it is available. 

The change effectively reverts the setup.py changes in dd579cd and b95ac55, although that the 'detailed instructions' section introduced in dd579cd becomes relevant. In fact the option to install with `--no-binary` is only meaningful when the default requirement is the psycopg2-binary.

I think that this is a reasonable proposal, but would like to know if there were any other reasons behind b95ac55 that would prevent using psycopg2 2.8 being usable.