Stop passing around the cli object so much.
Stdout and Pager can just be the same thing "OutputHandler" that just
wraps a function which handles output, this way the CLI can create
instances for each by passing `click.echo` and `self.echo_via_pager` and
the output module avoids unnecessary knowledge/coupling.

Log is a little trickier to de-couple from CLI, because it can't just
wrap a file path (the only thing it knew about on CLI) since that gets
changed all the time mid-process. So it grows a method for updating it's
target path and CLI can use that instead of changing `self.output_file`.